+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## v2.0.24
+
+- Adds an alternative to props for the Modal component through the use of slots
+
 ## v2.0.23
 
 - Adds `BankAccount` to Icons and the `Icon` component